backend/requirements.txt
flask==3.1.3
flask-cors==6.0.2
scikit-learn==1.7.2
numpy==2.1.3
requests==2.32.5
gunicorn==21.2.0
python-whois==0.9.6
